validacion de editar alumno. cancelar, activar evaluacion. envio de correos en evaluaciones y al cambiar el correo del alumno

+++ app/Http/Controllers/Tutorship/Tutstudent/TutstudentController.php
@@ -2,6 +2,7 @@
 
 namespace Intranet\Http\Controllers\Tutorship\Tutstudent;
 
+use Mail;
 use Illuminate\Http\Request;
 use Intranet\Http\Requests;
 use Intranet\Http\Requests\TutstudentRequest;
@@ -170,28 +171,53 @@ class TutstudentController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, $id)
+    public function update(TutstudentRequest $request, $id)
     {
-        try {            
-            //se busca un alumno con el mismo codigo
-            $u = User::where('Usuario',$request['codigo'])->first();
-            if($u!=null){
-                return redirect()->route('alumno.create')->with('warning', 'El código de alumno que se intenta registrar ya existe.');
-            }
-
+        try {   
             $student = Tutstudent::find($id);
-            $user = User::find($student->id_usuario);
-            //cambio el usuario del alumno
-            $user->Usuario = $request['codigo'];
-            $user->save();
 
-            //cambio el alumno            
-            $student->codigo       = $request['codigo'];            
+            //si cambio el codigo del alumno
+            if($request['codigo'] != $student->codigo){
+                //se busca un alumno con el mismo codigo
+                $u = User::where('Usuario',$request['codigo'])->first();
+                if($u!=null){
+                    return redirect()->route('alumno.create')->with('warning', 'El código de alumno que se intenta registrar ya existe.');
+                }
+
+                $user = User::find($student->id_usuario);
+                //cambio el usuario del alumno
+                $user->Usuario = $request['codigo'];
+                $user->save();
+                $student->codigo       = $request['codigo'];            
+            }
+            
+            
+
+            //cambio el alumno    
             $student->nombre       = $request['nombre'];            
             $student->ape_paterno  = $request['app'];            
-            $student->ape_materno  = $request['apm'];            
-            $student->correo       = $request['correo'];
-            $student->save();
+            $student->ape_materno  = $request['apm'];  
+
+            //si cambio el correo
+            if($request['correo'] != $student->correo )   {
+                $student->correo       = $request['correo'];  
+                //envio correo a la nueva direccion
+                try{
+                    $nombre = $request['nombre'];
+                    $mail = $request['correo'];
+                    Mail::send('emails.changeEmail',compact('nombre'),  function($m) use($mail) {
+                        $m->subject('Correo modificado');
+                        $m->to($mail);
+                    });
+                }
+                catch (\Exception $e)
+                {
+                    dd($e->getMessage());
+                }  
+            }       
+            
+            $student->save();         
+
             return redirect()->route('alumno.index', $id)->with('success', 'El alumno se ha actualizado exitosamente');
         } catch (Exception $e) {
             return redirect()->back()->with('warning', 'Ocurrió un error al hacer esta acción');
@@ -207,16 +233,14 @@ class TutstudentController extends Controller
     public function destroy($id)
     {
         try {
-            $student   = Tutstudent::find($id);
-            $message = "";
-            // if(count($area->investigators)){
-            //     return redirect()->back()->with('warning', 'Esta area esta asignada a investigadores');
-            // }
+            $student   = Tutstudent::find($id);            
             $student->id_tutoria = null;
             $student->save();
             
             $student->delete();
-            $student->tutorship->delete();
+            if($student->tutorship){
+                $student->tutorship->delete();    
+            }
 
             return redirect()->route('alumno.index')->with('success', 'El alumno se ha desactivado exitosamente');
         } catch (Exception $e) {
